requirements.txt
Flask
pandas
numpy
torch
scikit-learn
groq
chromadb
sentence-transformers
python-dotenv

